requirements.txt
# Core
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# API layer
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0

# LLM providers
openai>=1.30.0
anthropic>=0.25.0

# Local LLM (optional — for internet-off mode)
# ollama>=0.1.0

# Data connectors
pyathena>=3.0.0
sqlalchemy>=2.0.0

# Ingestion
pdfplumber>=0.9.0
pypdf>=3.0.0
python-docx>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
openpyxl>=3.1.0

# Forecasting
prophet>=1.1.0

# Config + utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Optional enrichment
# kafka-python>=2.0.0
# umap-learn>=0.5.0
